Move execution tests into flowr
