--- flowc/tests/test-flows/duplicate-connection/expected.stdout
@@ -1,2 +1,0 @@
-1
-1
