--- flowc/tests/test-flows/duplicate-connection/root.toml
@@ -1,17 +1,0 @@
-flow = "duplicate-connection"
-
-[[process]]
-alias = "args"
-source = "context://args/get"
-
-[[process]]
-alias = "print"
-source = "context://stdio/stdout"
-
-[[connection]]
-from = "args/string/1"
-to = "print"
-
-[[connection]]
-from = "args/string/1"
-to = "print"
