+++ flowr/tests/test-flows/args_json/expected.stdout
@@ -1,0 +1,1 @@
+3
